Moved message to separate class to avoid inheritance
