+++ src/Everscale/CommonTON/Messages.h
@@ -1,0 +1,77 @@
+// Copyright © 2017-2022 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
+#pragma once
+
+#include <optional>
+#include <utility>
+
+#include "Address.h"
+#include "../CellBuilder.h"
+#include "../CellSlice.h"
+#include "../Wallet.h"
+
+#include "../../PrivateKey.h"
+
+namespace TW::CommonTON {
+using CellBuilder = Everscale::CellBuilder;
+using StateInit = Everscale::StateInit;
+using Cell = Everscale::Cell;
+
+using uint128_t = CellBuilder::uint128_t;
+
+class CommonMsgInfo {
+public:
+    virtual void writeTo(CellBuilder& builder) const = 0;
+
+    virtual ~CommonMsgInfo() = default;
+};
+
+class ExternalInboundMessageHeader : public CommonMsgInfo {
+    AddressData _dst;
+    uint128_t _importFee{};
+
+public:
+    explicit ExternalInboundMessageHeader(AddressData dst)
+        : _dst(dst) {}
+
+    void writeTo(CellBuilder& builder) const override;
+};
+
+class InternalMessageHeader : public CommonMsgInfo {
+    bool _ihrDisabled;
+    bool _bounce;
+    AddressData _dst;
+    uint128_t _value;
+
+    bool _bounced{};
+    uint128_t _ihrFee{};
+    uint128_t _fwdFee{};
+    uint64_t _createdLt{};
+    uint32_t _createdAt{};
+
+public:
+    InternalMessageHeader(bool ihrDisabled, bool bounce, AddressData dst, uint64_t value)
+        : _ihrDisabled(ihrDisabled), _bounce(bounce), _dst(dst), _value(static_cast<uint128_t>(value)) {}
+
+    void writeTo(CellBuilder& builder) const override;
+};
+
+struct MessageData {
+    std::shared_ptr<CommonMsgInfo> header;
+    std::optional<StateInit> init{};
+    std::optional<Cell::Ref> body{};
+
+    using HeaderRef = std::shared_ptr<CommonMsgInfo>;
+
+
+    explicit MessageData(HeaderRef header) : header(std::move(header)) { }
+};
+
+MessageData createSignedMessage(PublicKey& publicKey, PrivateKey& key, bool bounce, uint32_t flags, uint64_t amount,
+                         uint32_t expiredAt, AddressData destination, const Cell::Ref& contractData);
+
+} // namespace TW::CommonTON
