+++ src/Everscale/Messages.h
@@ -6,77 +6,23 @@
 
 #pragma once
 
-#include <optional>
-#include <utility>
-
-#include "Address.h"
-#include "CellBuilder.h"
-#include "CellSlice.h"
-#include "Wallet.h"
-
-#include "../PrivateKey.h"
+#include "CommonTON/Messages.h"
 
 namespace TW::Everscale {
 
-using uint128_t = CellBuilder::uint128_t;
-
-class CommonMsgInfo {
-public:
-    virtual void writeTo(CellBuilder& builder) const = 0;
-
-    virtual ~CommonMsgInfo() = default;
-};
-
-class ExternalInboundMessageHeader : public CommonMsgInfo {
-    AddressData _dst;
-    uint128_t _importFee{};
-
-public:
-    explicit ExternalInboundMessageHeader(AddressData dst)
-        : _dst(dst) {}
-
-    void writeTo(CellBuilder& builder) const override;
-};
-
-class InternalMessageHeader : public CommonMsgInfo {
-    bool _ihrDisabled;
-    bool _bounce;
-    AddressData _dst;
-    uint128_t _value;
-
-    bool _bounced{};
-    uint128_t _ihrFee{};
-    uint128_t _fwdFee{};
-    uint64_t _createdLt{};
-    uint32_t _createdAt{};
-
-public:
-    InternalMessageHeader(bool ihrDisabled, bool bounce, AddressData dst, uint64_t value)
-        : _ihrDisabled(ihrDisabled), _bounce(bounce), _dst(dst), _value(static_cast<uint128_t>(value)) {}
-
-    void writeTo(CellBuilder& builder) const override;
-};
+using MessageData = CommonTON::MessageData;
+using InternalMessageHeader = CommonTON::InternalMessageHeader;
 
 class Message {
-protected:
-    std::shared_ptr<CommonMsgInfo> _header;
-
-    std::optional<StateInit> _init{};
-    std::optional<Cell::Ref> _body{};
-
+private:
+    MessageData _messageData;
 public:
-    virtual ~Message() noexcept = default;
-    using HeaderRef = std::shared_ptr<CommonMsgInfo>;
-
-    explicit Message(HeaderRef header)
-        : _header(std::move(header)) {}
-
-    [[nodiscard]] virtual Cell::Ref intoCell() const;
-    inline void setBody(const Cell::Ref& body) { _body = body; }
-    inline void setStateInit(const StateInit& stateInit) { _init = stateInit; }
+    explicit Message(const MessageData& messageData) : _messageData(messageData) {}
+    [[nodiscard]] Cell::Ref intoCell() const;
+    inline void setBody(const Cell::Ref& body) { _messageData.body = body; }
+    inline void setStateInit(const StateInit& stateInit) { _messageData.init = stateInit; }
 };
 
-Data createSignedMessage(PublicKey& publicKey, PrivateKey& key, bool bounce, uint32_t flags, uint64_t amount,
-                         uint32_t expiredAt, Address destination, const Cell::Ref& contractData);
+static auto createSignedMessage = CommonTON::createSignedMessage;
 
 } // namespace TW::Everscale
